+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.2.0] - 2026-08-11
+
+### Added
+
+- Added a Workstation tab!
+    - This can be used to filter by valid recipes in the currently open workstation.
+- Added support for regex in fluid/effect matching.
+
+### Changed
+
+- Adjusted default search widget offset when aligned under the index tab.
+- Improved stack group parsing to be more lenient.
+
+### Fixed
+
+- Fixed issues with JEED.
+
 ## [4.1.1] - 2026-08-07
 
 ### Fixed